Use a different build backend for packaging
Now the project uses `scikit-build-core>=0.5.0` as its building backend.
The `setup.py` is no longer needed, as the new backend integrates with
CMake nicely. Few things to discuss before further changes apply:

- The backend does not allow reading dependencies from requirements.txt.
  The project maintainer needs to decide whether we need to maintain two
  sets of requirements at the same time.
- There might be differences in package METAINFO and other information,
  due to the two backends are not compatible with each other. How to
  tackle this?

And this PR is therefore not complete. If the maintainer decides to
remove requirements.txt then other things such as Github workflow will
make changes accordingly. Combining those changes in one PR for easier
navigation in the future. If the maintainer decides to maintain all
information the original package maintains, then we might as rollback to
build script.

Signed-off-by: gxu <georgexu420@gmail.com>