Update CMake to work with new build backend
We no longer need to hard code (well, still a little bit) the version
number in CMakeLists.txt now. When invoked by the python build backend,
the backend automatically define some variables for us. We use them.

Also do not build tests when we are not manually running CMake. This is
temporary, and will be changed once we get real unit tests.

Signed-off-by: gxu <georgexu420@gmail.com>