Implementation of the object storage server protocol v5
For now, blocking happens on the level of requests instead of the level
of connections. This means every requests gets an id, and every response
contains an id, where these two ids are the same.

Get request when no object presents will be blocked. However, client is
able to send other requests (or even another blocking get request), and
get corresponding response back from the same connection channel.

Notice that the order of responses is not guaranteed anymore; Should
user wish to have an order, they need to sort responses based on the
responseId, with the same relation they have in the requestId.

Signed-off-by: gxu <georgexu420@gmail.com>