Change protocol (#102)
* Change protocol

Signed-off-by: gxu <georgexu420@gmail.com>

* Implementation of the object storage server protocol v5

For now, blocking happens on the level of requests instead of the level
of connections. This means every requests gets an id, and every response
contains an id, where these two ids are the same.

Get request when no object presents will be blocked. However, client is
able to send other requests (or even another blocking get request), and
get corresponding response back from the same connection channel.

Notice that the order of responses is not guaranteed anymore; Should
user wish to have an order, they need to sort responses based on the
responseId, with the same relation they have in the requestId.

Signed-off-by: gxu <georgexu420@gmail.com>

* Change printf-to-stdout to cerr

Signed-off-by: gxu <georgexu420@gmail.com>

* Make sure the pipeline don't fail because of upstream change

The pipeline failed because of yfinance's code. It uses method that are
not allowed in Python 3.8 when shipping this wheel. We use an older
version for now to avoid the hassle.

Test your code before you ship it to like, 3 mill people?

Signed-off-by: gxu <georgexu420@gmail.com>

* Solved a bug

Signed-off-by: gxu <georgexu420@gmail.com>

* Rename class variable with CamelCase

Signed-off-by: gxu <georgexu420@gmail.com>

* Use fail fast strategy.

For now we just add an explanation and terminate, maybe there are more
interesting things to do for asio errors. CAPNP provide pretty good
exception error messages, for example a stack dump, to stderr. So we are
most likely in a good position.

Signed-off-by: gxu <georgexu420@gmail.com>

---------

Signed-off-by: gxu <georgexu420@gmail.com>
Co-authored-by: gxu <georgexu420@gmail.com>