Initial Implementation of DenseUniqueValueMap
DenseUniqueValueMap is used to replace the OneToManyDict that is being
used in `ClientManager`. Given that `ClientManager` is the only user of
the original data structure, this commit aims to provide an alternative
to OneToManyDict that is less generic but more compact in terms of RAM
usage.

DenseUniqueValueMap does not provide all the APIs that OneToManyDict
provides. Instead, it provides only the API that is being used by the
user (aka. `ClientManager`). There are 4 APIs in total.

Current implementation of DenseUniqueValueMap is not final. There are
many things missing, for example proper `typedef`s are needed.

NOTE: There is another API that is being used in `ClientManager`. That
is `self._client_to_task_ids.get_values(client)`. The caller to this
API, however, is not being used in scaler. Perhaps it should be remove
later.

Signed-off-by: gxu <georgexu420@gmail.com>